refactor: fetcher 함수 위치 조정
